extract excerpt iframe to separate component
then we can use it when setting start/stop points in "add excerpt to drawer" button
